pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-eval-kit"
version = "0.1.0"
description = "Eval framework for LLM outputs — quality scoring, golden test suites, regression detection, cost tracking."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Akshay Dixit"},
]
keywords = ["llm", "eval", "evaluation", "quality", "testing", "prompt", "ai", "hallucination"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pyyaml>=6.0",
    "click>=8.1.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
llm = [
    "openai>=1.0",
    "anthropic>=0.18",
    "tiktoken>=0.5.0",
]
all = ["llm-eval-kit[llm]"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.3.0",
    "mypy>=1.8",
]

[project.scripts]
llm-eval = "llm_eval_kit.cli:main"

[project.urls]
Homepage = "https://github.com/kustonaut/llm-eval-kit"
Repository = "https://github.com/kustonaut/llm-eval-kit"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
